+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/webhook/DeleteWebhook.kt
@@ -1,0 +1,13 @@
+package com.github.insanusmokrassar.TelegramBotAPI.requests.webhook
+
+import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.internal.BooleanSerializer
+
+@Serializable
+class DeleteWebhook : SimpleRequest<Boolean> {
+    override fun method(): String = "deleteWebhook"
+
+    override fun resultSerializer(): KSerializer<Boolean> = BooleanSerializer
+}